filter delete bahan

+++ application/models/Bahan_model.php
@@ -12,9 +12,11 @@
             return $this->db->query("SELECT kode_bahan FROM bahan ORDER BY kode_bahan DESC LIMIT 1");
         }
         function delete_bahan($data){
-            // $count = $this->db->query("SELECT count(id_bahan) FROM bahan");
-
-            return $this->db->delete('bahan', $data);
+            // $count = $this->db->query("SELECT count(id_bahan) FROM resep WHERE id_bahan = $data");
+            $query = $this->db->get_where('resep', $data);
+            if($query->num_rows()==0){
+                return $this->db->delete('bahan', $data);
+            }
         }
     }
 ?>
